# cleanup and optimize rrdeng_load_metric_next()